flask-api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
SQLAlchemy==2.0.23
GeoAlchemy2==0.14.2
Shapely==2.0.2
PyMySQL==1.1.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.2
click==8.1.7
Flask-Caching==2.1.0
redis==5.0.1
Flask-JWT-Extended==4.6.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
flake8==6.1.0
gunicorn==21.2.0
